+++ Library/KLIMBinaryImage.m
@@ -28,9 +28,9 @@
         _buffer = malloc(_width * _height);
         UInt8 *bufferPtr = _buffer;
         
-        for (int y = 0; y < _height; y++) {
+        for (NSInteger y = 0; y < _height; y++) {
             UInt8 *pixelPtr = imagePtr + y * bytesPerRow;
-            for (int x = 0; x < _width; x++) {
+            for (NSInteger x = 0; x < _width; x++) {
                 // 色情報を取得する
                 UInt8 b = *pixelPtr++;  // 青
                 UInt8 g = *pixelPtr++;  // 緑
@@ -66,9 +66,9 @@
         UInt32 *accumBuffer = malloc(_width * _height * 4);
         UInt32 *accumPtr = accumBuffer;
         
-        for (int y = 0; y < _height; y++) {
+        for (NSInteger y = 0; y < _height; y++) {
             UInt8 *pixelPtr = imagePtr + y * bytesPerRow;
-            for (int x = 0; x < _width; x++) {
+            for (NSInteger x = 0; x < _width; x++) {
                 // 色情報を取得する
                 UInt8 b = *pixelPtr++;  // 青
                 UInt8 g = *pixelPtr++;  // 緑
@@ -91,15 +91,15 @@
         CFRelease(dataRef);
         
         bufferPtr = _buffer;
-        int R = _width / 8;
-        for (int y = 0; y < _height; y++) {
-            int ymin = MAX(0, y - R);
-            int ymax = MIN(_height - 1, y + R);
-            int iymin = ymin ? (ymin - 1) * _width : 0;
-            int iymax = ymax * _width;
-            for (int x = 0; x < _width; x++) {
-                int xmin = MAX(0, x - R);
-                int xmax = MIN(_width - 1, x + R);
+        NSInteger R = _width / 8;
+        for (NSInteger y = 0; y < _height; y++) {
+            NSInteger ymin = MAX(0, y - R);
+            NSInteger ymax = MIN(_height - 1, y + R);
+            NSInteger iymin = ymin ? (ymin - 1) * _width : 0;
+            NSInteger iymax = ymax * _width;
+            for (NSInteger x = 0; x < _width; x++) {
+                NSInteger xmin = MAX(0, x - R);
+                NSInteger xmax = MIN(_width - 1, x + R);
                 UInt32 sum = accumBuffer[iymax + xmax];
                 if (xmin) {
                     sum -= accumBuffer[iymax + xmin];
@@ -110,7 +110,7 @@
                         sum += accumBuffer[iymin + xmin];
                     }
                 }
-                int val = *bufferPtr * (xmax - xmin + 1) * (ymax - ymin + 1);
+                NSInteger val = *bufferPtr * (xmax - xmin + 1) * (ymax - ymin + 1);
                 sum *= 0.75;
                 UInt8 p = val < sum ? 255 : 0;
                 *bufferPtr++ = p;
@@ -129,7 +129,7 @@
         _height = image.height;
         
         // int bufferSize = _width * _height * 4;
-        int bufferSize = _width * _height;
+        NSInteger bufferSize = _width * _height;
         _buffer = malloc(bufferSize);
         
         memcpy(_buffer, image.buffer, bufferSize);
@@ -146,16 +146,16 @@
 
 - (void)contract
 {
-    int bufferSize = _width * _height;
+    NSInteger bufferSize = _width * _height;
     UInt8 *buffer = malloc(bufferSize);
     
     memcpy(buffer, _buffer, bufferSize);
     
     //収縮
-    for(int y = 2; y < _height - 2; y++) {
-        for(int x = 2; x < _width - 2; x++) {
-            int c = y * _width + x;
-            int b = 1;
+    for(NSInteger y = 2; y < _height - 2; y++) {
+        for(NSInteger x = 2; x < _width - 2; x++) {
+            NSInteger c = y * _width + x;
+            NSInteger b = 1;
             if (!buffer[c]) {
                 b = 0;
             } else {
@@ -174,16 +174,16 @@
 
 - (void)expand
 {
-    int bufferSize = _width * _height;
+    NSInteger bufferSize = _width * _height;
     UInt8 *buffer = malloc(bufferSize);
         
     memcpy(buffer, _buffer, bufferSize);
     
     //膨張
-    for(int y = 2; y < _height - 2; y++) {
-        for(int x = 2; x < _width - 2; x++) {
-            int c = y * _width + x;
-            int b = 0;
+    for(NSInteger y = 2; y < _height - 2; y++) {
+        for(NSInteger x = 2; x < _width - 2; x++) {
+            NSInteger c = y * _width + x;
+            NSInteger b = 0;
             if (buffer[c]) {
                 b = 1;
             } else {
@@ -202,12 +202,12 @@
 
 - (void)ensureClearEdge
 {
-    int y = _height - 1;
-    for (int x = 0; x < _width; x++) {
+    NSInteger y = _height - 1;
+    for (NSInteger x = 0; x < _width; x++) {
         _buffer[x] = 0;
         _buffer[y * _width + x] = 0;
     }
-    for (int y = 1; y < _height - 1; y++) {
+    for (NSInteger y = 1; y < _height - 1; y++) {
         _buffer[(y - 1) * _width + 1] = 0;
         _buffer[y * _width] = 0;
     }
@@ -215,15 +215,15 @@
 
 - (void)removeCrumb
 {
-    int bufferSize = _width * _height;
+    NSInteger bufferSize = _width * _height;
     UInt8 *buffer = malloc(bufferSize);
     
     memcpy(buffer, _buffer, bufferSize);
     
-    for(int y = 1; y < _height - 1; y++) {
-        for(int x = 1; x < _width - 1; x++) {
-            int c = y * _width + x;
-            int b = buffer[c];
+    for(NSInteger y = 1; y < _height - 1; y++) {
+        for(NSInteger x = 1; x < _width - 1; x++) {
+            NSInteger c = y * _width + x;
+            NSInteger b = buffer[c];
             if (b) {
                 if (!(buffer[c+1] || buffer[c-1] || buffer[c+_width] || buffer[c-_width] ||
                       buffer[c+1+_width] || buffer[c+1-_width] ||
@@ -250,8 +250,8 @@
     // UInt32 *bufferPtr = self.buffer;
     UInt8 *bufferPtr = self.buffer;
     
-    for (int y = 0; y < _height; y++) {
-        for (int x = 0; x < _width; x++) {
+    for (NSInteger y = 0; y < _height; y++) {
+        for (NSInteger x = 0; x < _width; x++) {
             
             // ピクセルのポインタを取得する
             UInt8* pixelPtr = imagePtr + (y * self.width + x) * 4;
